fix(onboarding): replace JWT gate with server-component DB read
PR #99 added a middleware-level onboarding gate keyed on
token.hasOnboarded. The flag was refreshed via session.update() after
onboarding submitted (PR #101 added a hard reload to dodge a related
cookie-commit race). But the founder still hit the same bug in
manual testing: complete onboarding, get bounced back to /onboarding,
form re-mounts empty. Signout/signin fixes it.

Root cause: NextAuth's session.updateAge defaults to 1 day. Inside
that window NextAuth doesn't re-issue the JWT cookie even when the
jwt callback returns a token with updated fields. Our updateSession()
calls ran the callback, set the new hasOnboarded value in memory, and
sent a response — but the cookie wasn't re-set, so the next request
still had hasOnboarded=false. Sign-out + sign-in works because it
issues a fresh JWT, bypassing updateAge.

This is the second time the JWT-staleness class of bug has bit us
(PR #94 was the first, for credits-after-onboarding). The pattern:
anything that bridges a DB write to a JWT claim needs explicit
session refresh in every mutating call site, and one missed call
silently breaks the gate. We've been chasing this once per PR.

Switch to a server-component DB read. The dashboard layout (new
nested file at (dashboard)/dashboard/layout.tsx) calls a
requireOnboarded() helper that runs auth() + findUnique(
organizationName) on every server-rendered request. The onboarding
page itself does the mirror check (requireNotOnboarded). One source
of truth — the DB — read fresh every server render. No staleness.

Cost: a single indexed PK lookup per server-rendered dashboard
visit. Client-side navigations within the dashboard reuse the
layout and skip the check, so the read fires on first load, hard
refresh, deep link — single digit times per session.

Changes:

- New src/lib/onboarding-guard.ts: getOnboardingStatus(),
  requireOnboarded(), requireNotOnboarded(). Single helper module so
  the rules live in one place.
- New src/app/(dashboard)/dashboard/layout.tsx: thin server-component
  wrapper calling requireOnboarded(). Scopes the gate to /dashboard/*
  by directory placement, leaving /onboarding (which lives in
  (dashboard)/onboarding) untouched.
- src/app/(dashboard)/onboarding/page.tsx is now a server component
  that calls requireNotOnboarded() and renders the form. The form
  itself moved to ./onboarding-form.tsx (client component). The form
  no longer calls useSession().update() — it just hard-navigates to
  /dashboard and trusts the layout guard to confirm.
- src/lib/auth.ts: drop the hasOnboarded JWT plumbing (initial-signin
  set, OAuth re-read, trigger=update re-read). Credentials authorize()
  no longer returns organizationName. JWT is back to its pre-#99 shape.
- src/middleware.ts: drop both onboarding-gate branches (forward and
  mirror). Middleware is back to auth + signin-redirect only.
- src/types/next-auth.d.ts: drop hasOnboarded from Session and JWT.

Tests:

- tests/unit/lib/onboarding-guard.test.ts (NEW): 10 cases covering
  getOnboardingStatus, requireOnboarded (redirect/passthrough/anon),
  requireNotOnboarded (redirect/passthrough/anon). next/navigation
  redirect() is mocked as a sentinel-throwing function.
- tests/unit/middleware.test.ts: rewritten. The 3 onboarding-gate
  cases are gone (now covered by the guard test). The remaining 5
  cases (signin redirect, 401 for unauthenticated API, founder
  exemption, 404 for non-founder admin, /pricing passthrough)
  stay, plus 3 new cases asserting that middleware no longer does
  any onboarding-gate work.

Total suite: 722 passing (up from 712). No regressions.
